#125 fix clang compilation -Wold-style-cast

+++ dynawo/sources/Models/CPP/Events/ModelVoltageSetPointChange/DYNModelVoltageSetPointChange.cpp
@@ -133,7 +133,7 @@ ModelVoltageSetPointChange::setGequations() {
   gEquationIndex_[0] = "stopTime > t >= startTime";
   gEquationIndex_[1] = "t >= stopTime";
 
-  assert(gEquationIndex_.size() == (unsigned int) sizeG() && "Model VoltageSetPointChange: gEquationIndex.size() != gLocal_.size()");
+  assert(gEquationIndex_.size() == static_cast<size_t>(sizeG()) && "Model VoltageSetPointChange: gEquationIndex.size() != gLocal_.size()");
 }
 
 void
